requirements.txt
streamlit==1.24.1
google-generativeai==0.8.0  # Or the appropriate version
